requirements.txt
requests>=2.28
openpyxl>=3.1
pyyaml>=6.0
